Rename model-v2 and associated code to just model
